[xlsb] fix max ccolspan (#1030)
* [xlsb] fix max ccolspan

* [misc] fix sign-compare warning

* [xlsb] fix reading long spans. This is not entirely correct. See the comment.

* [xlsb] fix text rotation output

+++ src/helper_functions.cpp
@@ -487,7 +487,7 @@ void wide_to_long(
   R_xlen_t idx = 0;
 
   bool has_dims = false;
-  if (dims.size() == n * m) has_dims = true;
+  if (dims.size() == (size_t)(n * m)) has_dims = true;
 
   for (auto i = 0; i < m; ++i) {
     Rcpp::checkUserInterrupt();
